+++ src/Listeners/LaravelMixpanelUserObserver.php
@@ -23,11 +23,11 @@ class LaravelMixpanelUserObserver
      */
     public function created(Model $user)
     {
-        $firstName = $user->first_name ?: '';
-        $lastName = $user->last_name ?: '';
+        $firstName = $user->getAttribute('first_name');
+        $lastName = $user->getAttribute('last_name');
 
-        if ($user->name) {
-            $nameParts = explode(' ', $user->name);
+        if ($user->getAttribute('name')) {
+            $nameParts = explode(' ', $user->getAttribute('name'));
             array_filter($nameParts);
             $lastName = array_pop($nameParts);
             $firstName = implode(' ', $nameParts);
@@ -36,13 +36,12 @@ class LaravelMixpanelUserObserver
         $data = [
             '$first_name' => $firstName,
             '$last_name' => $lastName,
-            '$name' => $user->name,
-            '$email' => $user->email,
+            '$name' => $user->getAttribute('name'),
+            '$email' => $user->getAttribute('email'),
+            '$created' => ($user->getAttribute('created_at')
+                ? $user->getAttribute('created_at')->format('Y-m-d\Th:i:s')
+                : null),
         ];
-
-        if ($user->created_at) {
-            $data['$created'] = $user->created_at->format('Y-m-d\Th:i:s');
-        }
 
         array_filter($data);
 
@@ -57,11 +56,11 @@ class LaravelMixpanelUserObserver
     public function saving(Model $user)
     {
         $this->mixPanel->identify($user->getKey());
-        $firstName = $user->first_name ?: '';
-        $lastName = $user->last_name ?: '';
+        $firstName = $user->getAttribute('first_name');
+        $lastName = $user->getAttribute('last_name');
 
-        if ($user->name) {
-            $nameParts = explode(' ', $user->name);
+        if ($user->getAttribute('name')) {
+            $nameParts = explode(' ', $user->getAttribute('name'));
             array_filter($nameParts);
             $lastName = array_pop($nameParts);
             $firstName = implode(' ', $nameParts);
@@ -70,13 +69,12 @@ class LaravelMixpanelUserObserver
         $data = [
             '$first_name' => $firstName,
             '$last_name' => $lastName,
-            '$name' => $user->name,
-            '$email' => $user->email,
+            '$name' => $user->getAttribute('name'),
+            '$email' => $user->getAttribute('email'),
+            '$created' => ($user->getAttribute('created_at')
+                ? $user->getAttribute('created_at')->format('Y-m-d\Th:i:s')
+                : null),
         ];
-
-        if ($user->created_at) {
-            $data['$created'] = $user->created_at->format('Y-m-d\Th:i:s');
-        }
 
         array_filter($data);